Improve os.pwd: Allow to provide custom "current working directory" supplier
Java has no notion of a current working directory. The often used
idiom

	Paths.get("").toAbsolutePath()

only returns the *initial* working directory and does not change over
the processes lifetime, even if the process calls chdir().

Until this commit, os.pwd uses this idiom. This means that users who
need a reliable way to the the *current* working directory of a
process could not use os.pwd. This commit introduces

	os.setCwdSupplier(supplier: () => Path)

that is, a method which allows users to provide their own CWD
supplier. A typical candidate would be

    os.setCwdSupplier(() =>
		os.Path(jnr.posix.POSIXFactory.getPOSIX().getcwd())

from jnr-posix.