Fix os.pwd: evalute every time and use portable approach
The current working directory of a process may change over time, hence
os.pwd must not be a value that is initialized once, but instead a
callable function.

Furthermore, I think the more portable approach to get the current
working directory is

Paths.get("").toAbsolutePath()

as "." appears to be Unix specific.