Add Path.of(PathConvertible)
OS-Lib was lacking a convinience methods that lets user construct
absolute paths from a wide range of inputs values. In particular,
there is no function that takes either an absolute or relativ path
input and constructs an absolute os.Path from it.

The newly added Path.of() is exactly such a function, plus it includes
the functionality of the already existing Path.expandUser() function.