Add the actual command to CommandResult
The messages of exceptions thrown by call() would previously just
mention the exit status of the invoked process, but not the actual
command line that was used to create the process. To help with
debugging purposes, the exception message now includes the full
command line and the os.proc instance provides access to it.