Add os.Path.of()
OS-Lib was lacking a convenience methods that lets user construct
absolute paths from a wide range of inputs values. In particular,
there is no function that takes either an absolute path, relative path
or a tilde prefixed path and constructs an absolute os.Path from it.

The newly added Path.of() is exactly such a function, plus it includes
the functionality of the already existing Path.expandUser() function.