Adjust os.Path.apply() to take relative paths and expand home
OS-Lib was lacking a convenience methods that lets user construct
absolute paths from a wide range of inputs values. In particular,
there is no function that takes either an absolute path, relative path
or a tilde prefixed path and constructs an absolute os.Path from
it.

The new behavior of os.Path.apply() accommodates for this.