fix: handle `baseName` for hidden dirs
I had a usecase for this when I was working with `.well-known` as a dir
and got bite by this since `baseName` on this will just return `""`.
This changes the logic slightly to return `last` if either `-1` is found
meaning there is no `.` in the file, or if the index is `0` indicating
that it's hidden and there is also no extension. I've also added a test
for this.