Add javadoc.io link and badge to Readme (#180)
While OS-Lib is small and the Readme has the cookbook
already, it'd also be helpful to have some API documentation to search
through. 

This adds a link to https://javadoc.io/doc/com.lihaoyi/os-lib_3 as well as a badge at the
top of the Readme. I only put in a link to the docs for the `os-lib_3`
artifact since I figured it'd be the same for Scala 3 as well as Scala
2.13, but if not, I guess two links, one for Scala 3 and one for Scala
2, can be put in.

Pull request: https://github.com/com-lihaoyi/os-lib/pull/180

Co-authored-by: Tobias Roeser <le.petit.fou@web.de>