Introduce `os.proc.env` `DynamicVariable` to override default `env`
In Mill and any client-server application, it is handy to start processes on the server passing the
client environment. This requires to pass the environment manually in all the `os.proc(...).call()`s.
An easier alternative is to override the environment as we to for `os.Inherit` `out`, `in` and `err` pipes.
So users can forget about the environment and have the right one passed automatically.