Make `os.pwd` modifiable via the `os.pwd0` dynamic variable (#298)
Basically the same thing we did in
https://github.com/com-lihaoyi/os-lib/pull/295 and
https://github.com/com-lihaoyi/os-lib/pull/283, except rather than being
`os.SubProcess`-specific this applies to all usage of `os.pwd`. And with
similar limitations: this works for code paths that go through OS-Lib,
but not for code that uses `java.io` or `java.nio` directly.

AFAIK this is the last of the "global" things we need to redirect in
Mill tasks, after env variables and default subprocess streams. I'd like
it to go into Mill 0.12.0 together with the other sandboxing-related
changes

Added unit tests and updated the documentation

Pull Request: https://github.com/com-lihaoyi/os-lib/pull/298