Allow even more dynamic instrumentation of `pwd` via `os.dynamicPwdFunction` (#299)
Follow up to https://github.com/com-lihaoyi/os-lib/pull/298, which it
turns out was not flexible enough to do what we need in Mill. Mill uses
references to the `pwd` to create the `.dest` folder lazily, thus we
cannot just assign a value to the `os.pwd` but actually need to assign a
lambda that can be replaced to implement the lazy evaluation

I pulled the trigger too quickly on releasing 0.10.6 with
`os.dynamicPwd`, so this one has to be named differently. I chose
`os.dynamicPwdFunction` and will make sure to test it with an unstable
release before releasing a stable tag