Make use of concise literal paths everywhere (#302)
Both test suite and docs. We leave those in `PathTests` unchanged, since
that file intentionally contains a mix of different concise and verbose
paths for testing purposes, and we also leave a handful of examples of
verbose paths in the docs to demonstrate it's possible. Otherwise
everywhere uses the concise path syntax