Add support for permissions and symlinks in os.zip/unzip with vendored zip source code from Apache Ant (#374)
### Vendored code

Follow the discussion in #356, this adds support for permissions and
symlinks in `os.zip`/`unzip` with vendored zip source code from Apache
Ant.

The vendored source code is generated by the `os.zip.apacheAntZipSource`
task and put in `os/zip`. It's shaded with the package renamed from
`org.apache.tools.zip` to `os.shaded_org_apache_tools_zip`.

`scala-steward.conf` was added and configured to run
`os.zip.apacheAntZipSource` on `org.apache.ant:ant` updates.

### Features

This brings support for permissions and symlinks to `zip` (for creating
new zips, not modifying existing ones), `zip.stream` and `unzip`. As for
modifying existing zips, we would still have to rely on `jdk.zipfs`
which does not support symlinks.

| | file permissions | symlinks |
| --- | --- | --- |
| `os.zip.open` | if Java Runtime Version >= 14 | |
| `os.zip` (create new) | ✅ | ✅ |
| `os.zip` (modify existing) | if Java Runtime Version >= 14 | |
| `os.zip.stream` | ✅ | ✅ |
| `os.unzip` | ✅ | ✅ |
| `os.unzip.stream` | | |

### TODO
- [ ] **(Advice needed)** make sure we comply with Apache Ant's license
to include the code here. Would appreciate opinions on this as I'm not
an expert.
- [ ] **(Advice needed)** make `ZipOps` JVM only
- [x] tests
- [x] make sure things don't break on Windows
- [x] add permission support to modifying existing zips with `jdk.zipfs`
like what @sake92 did in #371